+++ packages/types/src/metadata/v11/toV12.ts
@@ -11,13 +11,9 @@ export function toV12 (registry: Registry, { extrinsic, modules }: MetadataV11):
   return registry.createTypeUnsafe('MetadataV12', [{
     extrinsic,
     modules: modules.map((mod): ModuleMetadataV12 =>
-      // ensure that we unwrap all options (conversion is problematic otherwise)
       registry.createTypeUnsafe('ModuleMetadataV12', [{
         ...mod,
-        calls: mod.calls.unwrapOr(null),
-        events: mod.events.unwrapOr(null),
-        index: 255,
-        storage: mod.storage.unwrapOr(null)
+        index: 255
       }])
     )
   }]);